add the most crucial car and offer functionalities

+++ src/main/java/pl/zzpj/autorent/autorent/services/CarService.java
@@ -22,37 +22,43 @@ public class CarService {
     public Car getCar(String id) {
         return carRepository.get(id).orElseThrow();
     }
-//
-//    // TODO: 03.05.2021 check car ownership
-//    // TODO: 03.05.2021 check if car is rented before upgrade
+
+    // TODO: 03.05.2021 check car ownership
     public void updateCar(String id, Car car) {
-        carRepository.update(id,car);
-        //return updatedCar;
+        if (!car.isRented()) {
+            carRepository.update(id, car);
+        }
     }
 
     public void rentCar(String id) {
         Car car = getCar(id);
         car.setRented(true);
-        carRepository.update(id,car);
+        carRepository.update(id, car);
     }
 
-//    // TODO: 03.05.2021 edit car
-//
     public void addCar(Car car) {
         car.setId(UUID.randomUUID().toString());
         carRepository.save(car);
     }
 
-//    // TODO: 03.05.2021 check car ownership
-//    // TODO: 03.05.2021 check if car is rented before deleting
+    // TODO: 03.05.2021 check car ownership
     public void deleteCar(String id) {
-        carRepository.deleteById(id);
+        Car car = getCar(id);
+        if (!car.isRented()) {
+            carRepository.deleteById(id);
+        }
     }
 
     public List<Car> getAllCars() {
         return carRepository.retrieveAll();
     }
-//
-//    // TODO: 03.05.2021 add method getAllNoRentedCar
-//
+
+    // TODO: 03.05.2021 add method getAllNoRentedCar
+
+    public List<Car> getAllNoRentedCars() {
+        List<Car> cars = carRepository.retrieveAll();
+        cars.removeIf(Car::isRented);
+        return cars;
+    }
+
 }
